Fixes and doc improvements (provides / requires was backwards)
